logging behavior

+++ eshop-microservices/src/Services/Catalog/Catalog.API/Products/DeleteProduct/DeleteProductHandler.cs
@@ -1,15 +1,28 @@
 ﻿
+using FluentValidation;
+
 namespace Catalog.API.Products.DeleteProduct
 {
-    public record DeleteProductCommand(int Id) : ICommand<DeleteProductResult> { }
+    public record DeleteProductCommand(Guid Id) : ICommand<DeleteProductResult> { }
     public record DeleteProductResult(bool deleted);
-    public class DeleteProductHandler(IDocumentSession session) : IRequestHandler<DeleteProductCommand, DeleteProductResult>
+    public class DeleteProductCommandValidator : AbstractValidator<DeleteProductCommand>
+    {
+        public DeleteProductCommandValidator()
+        {
+            RuleFor(x => x.Id).NotEmpty().WithMessage("Product ID is required");
+        }
+    }
+    internal class DeleteProductHandler(IDocumentSession session) : IRequestHandler<DeleteProductCommand, DeleteProductResult>
     {
         public async Task<DeleteProductResult> Handle(DeleteProductCommand command, CancellationToken cancellationToken)
         {
-                session.Delete<Product>(command.Id);
-                await session.SaveChangesAsync(cancellationToken);
-                return new DeleteProductResult(true);
+            var existedProduct = await session.LoadAsync<Product>(command.Id, cancellationToken);
+            if (existedProduct == null) {
+                throw new ProductNotFoundException(command.Id);
+            }
+            session.Delete<Product>(command.Id);
+            await session.SaveChangesAsync(cancellationToken);
+            return new DeleteProductResult(true);
 
         }
     }
